Fix credentials stored in Docker images
Fixes #55

When env vars are used to construct values in Next.js, a "const" declared at the
top level may be evaluated at build time, runtime by the server, or - in the
case of client components - runtime by the browser.

The PINECONE_API_KEY and PINECONE_ENVIRONMENT variables here need to be deferred
to runtime, and not used in static rendering. We put the constructor for
Pinecone behind a getter, which ensures `process.env.$KEY` is evaluated at
runtime.

We also mark the route as dynamic, this can be used by Next.js to defer
rendering of components that might call this route as well.